requirements.txt
flask>=3.0.0
pandas>=2.0.0
google-play-scraper>=1.2.6
requests>=2.31.0
python-dotenv>=1.0.0
